extract aggregationcursor and querycursor
